+++ components/index.ts
@@ -1,0 +1,5 @@
+export * from "./Button";
+export * from "./colors";
+export { default as Header } from "./Header";
+export { default as NavBar } from "./NavBar";
+export * from "./SpeakerCard";
